refactor(posts): restructure posts module for better maintainability
This completes Phase 1-3 of the refactoring plan: extracted types/schemas/utils to dedicated files, moved TypeSense operations to external dir, and extracted read operations to posts/read.ts. Reduced size of posts-query.ts and added JSDoc comments.
