+++ apps/egghead/src/lib/external/typesense/index.ts
@@ -1,0 +1,2 @@
+// Re-export everything from the TypeSense module
+export * from './post'
