+++ apps/egghead/src/lib/posts/index.ts
@@ -1,0 +1,11 @@
+// Re-export types
+export * from './types'
+
+// Re-export schemas
+export * from './schemas'
+
+// Re-export utility functions
+export * from './utils'
+
+// Note: We're not re-exporting from '../posts' here to avoid circular references
+// This file is intended to replace it eventually
