+++ apps/egghead/src/lib/posts/utils.ts
@@ -1,0 +1,24 @@
+import crypto from 'crypto'
+import { guid } from '@/utils/guid'
+import slugify from '@sindresorhus/slugify'
+
+import { Post } from './types'
+
+export function updatePostSlug(currentPost: Post, newTitle: string): string {
+	if (newTitle !== currentPost.fields.title) {
+		const splitSlug = currentPost?.fields.slug.split('~') || ['', guid()]
+		return `${slugify(newTitle)}~${splitSlug[1] || guid()}`
+	}
+	return currentPost.fields.slug
+}
+
+export function generateContentHash(post: Post): string {
+	const content = JSON.stringify({
+		title: post.fields.title,
+		body: post.fields.body,
+		description: post.fields.description,
+		slug: post.fields.slug,
+		// Add any other fields that should be considered for content changes
+	})
+	return crypto.createHash('sha256').update(content).digest('hex')
+}
